==> dino-cp-analysis/03-lincs-well-aggregation-sphering-vits.ipynb <==
import umap.umap_ as umap
from sklearn.preprocessing import StandardScaler
import torch
import numpy as np
from pathlib import Path
import matplotlib.pyplot as plt
import pandas as pd

import numpy as np
import sklearn.metrics
import os
import sys

from tqdm import tqdm

sys.path.append("../profiling/")
import profiling

feature_path = ''
feature_add_path = ''

output_folder = 'data'
output_file = "well_level_profiles_vits_LINCS_1e-5_final.csv"
REG_PARAM = 1e-5

# Load metadata
meta = pd.read_csv("Cell_Painting_data/LINCS_sc_metadata.csv")

files = [str(f) for f in Path(feature_path).glob('features*.pth')]
open_files = [torch.load(f) for f in files]
features = torch.concat([f[0] for f in open_files])

print("Total images:",features.shape[0])
print("Number of features per image:", features.shape[1])

scaled_features = StandardScaler().fit_transform(features.numpy())

cell_names = np.concatenate(([f[1] for f in open_files]))
order, ordered_features = (np.array(t) for t in zip(*sorted(zip(cell_names, scaled_features))))

# 2. Site-level profiles / Median Aggregation

group_dict = meta.groupby('Key').groups
print("Grouping finished.")

site_level_data = []
site_level_features = []

for site_name in tqdm(list(group_dict.keys())):
    metadata = site_name.split('/')
    indices = group_dict[site_name]
    mean_profile = np.median(ordered_features[indices], axis=0)
    
    site_level_data.append(
        {
            "Plate": metadata[0],
            "Well": metadata[1],
            "Treatment": meta["Treatment"][indices].unique()[0]
        }

    )
    site_level_features.append(mean_profile)


num_features = 384
columns1 = ["Plate", "Well", "Treatment"] # dataset
columns2 = [i for i in range(num_features)]

sites1 = pd.DataFrame(columns=columns1, data=site_level_data)
sites2 = pd.DataFrame(columns=columns2, data=site_level_features)
sites = pd.concat([sites1, sites2], axis=1)

sites["Treatment_Clean"] = sites["Treatment"].apply(lambda x: "-".join([str(i) for i in x.split("-")[:2]]))

# 3. Well-level profiles / Mean Aggregation

# Collapse well data
wells = sites.groupby(["Plate", "Well", "Treatment", "Treatment_Clean"]).mean().reset_index()
wells[:10]

wells.to_csv(f"{output_folder}/Wells_Prewhitened_ViT_small_LINCS.csv")

# 4. Whitening

sum(wells["Treatment"].isin(["DMSO@NA"]))

whN = profiling.WhiteningNormalizer(wells.loc[wells["Treatment"].isin(["DMSO@NA"]), columns2], REG_PARAM)
whD = whN.normalize(wells[columns2])

# Save whitened profiles
wells[columns2] = whD
wells.to_csv(f'{output_folder}/{output_file}', index=False)
