==> dino-cp-analysis/05-align-cellprofiler-profiles.ipynb <==
### - Split Compounds into Train & Test data based on the number of MOAs that are attributed to them.

This script is adapted from Gregory Way, Adeniyi Adeboye, & Shantanu Singh. (2022). broadinstitute/lincs-profiling-complementarity: Response to Reviewers (Version v2). Zenodo. https://doi.org/10.5281/zenodo.6522802

import os
import pathlib
import requests
import pickle
import argparse
import pandas as pd
import numpy as np
import re
from os import walk
from collections import Counter
import random
import shutil
from split_compounds import split_cpds_moas

## 1. Select max dose for Cell Profiler

# Copy well level profiles to '../source_data'
# Load Data file for Our Experiment
df_dino = pd.read_csv("data/well_level_profiles_vits_LINCS_1e-5_final.csv")
df_CNN = pd.read_csv('data/well_level_profiles_cpcnn_LINCS_1e-5_final.csv')
print(df_dino.shape, df_CNN.shape)

# Load CellProfiler's datafile
data_path = 'data/'

# Download cp_level4_cpd_replicates.csv.gz from https://github.com/broadinstitute/lincs-profiling-complementarity/blob/master/1.Data-exploration/Profiles_level4/cell_painting/cellpainting_lvl4_cpd_replicate_datasets/cp_level4_cpd_replicates.csv.gz
df_cellprofiler = pd.read_csv(
    os.path.join(data_path, 'cp_level4_cpd_replicates.csv.gz'), 
    compression='gzip',
    low_memory = False
)
print(df_cellprofiler.shape)

# Exclude DMSO
df_dino = df_dino[df_dino['Treatment'] != 'DMSO@NA'].reset_index(drop=True)

df_CNN = df_CNN[df_CNN['Treatment'] != 'DMSO@NA'].reset_index(drop=True)
df_CNN["Treatment_Clean"] = df_CNN["broad_sample"].apply(lambda x: '-'.join(x.split('-')[:2]))

df_cellprofiler = df_cellprofiler[df_cellprofiler['broad_id'] != 'DMSO'].reset_index(drop=True)

print(df_dino.shape, df_CNN.shape, df_cellprofiler.shape)

common_treatment = list(set(df_CNN["Treatment_Clean"].unique()) 
                        & set(df_cellprofiler["broad_id"].unique()))

len(common_treatment)

# Select rows with common treatments only
df_cellprofiler = df_cellprofiler.loc[df_cellprofiler['broad_id'].isin(common_treatment)]
df_dino = df_dino.loc[df_dino['Treatment_Clean'].isin(common_treatment)]
df_CNN = df_CNN.loc[df_CNN['Treatment_Clean'].isin(common_treatment)]

print(len(df_cellprofiler["broad_id"].unique()))
print(len(df_dino['Treatment_Clean'].unique()))
print(len(df_CNN["Treatment_Clean"].unique()))

# Filter for only max dose
idx = df_cellprofiler.groupby(['broad_id'])['Metadata_dose_recode'].transform(max) == \
        df_cellprofiler['Metadata_dose_recode']

df_cellprofiler = df_cellprofiler[idx]

print(df_cellprofiler.shape)
print(df_dino.shape)
print(df_CNN.shape)

## 2. Align MOA annotation

# Convert moa annotation to lower case
df_cellprofiler['moa'] = df_cellprofiler['moa'].apply(lambda x: x.lower())

# Create moa-compound dictionary
df_cpds_moas = df_cellprofiler.drop_duplicates(['broad_id','moa'])[['broad_id','moa']]
cpds_moa = dict(zip(df_cpds_moas['broad_id'], df_cpds_moas['moa']))
len(cpds_moa)

df_cpds_moas.to_csv('moa_annotation.csv', index=False)

# Concatenate moa for three datasets
df_dino["moa"]= df_dino["Treatment_Clean"].map(cpds_moa)
df_CNN['moa'] = df_CNN['Treatment_Clean'].map(cpds_moa)

print(len(df_cellprofiler["moa"].unique()), 
      len(df_dino['moa'].unique()), 
      len(df_CNN["moa"].unique()))

# Add compound name 'pert_iname' for dino and cpcnn features
pertname = df_cellprofiler.drop_duplicates(['pert_iname','broad_id'])[['pert_iname','broad_id']]
pertname_dict = dict(zip(pertname['broad_id'], pertname['pert_iname']))

df_dino['pert_iname'] = df_dino['Treatment_Clean'].map(pertname_dict)
df_CNN['pert_iname'] = df_CNN['Treatment_Clean'].map(pertname_dict)

#Save file to csv
out_dir = 'data'

df_cellprofiler.to_csv(f"{out_dir}/cp_cellprofiler_final.csv",index=False)
df_dino.to_csv(f"{out_dir}/cp_dino_final.csv",index=False)
df_CNN.to_csv(f"{out_dir}/cp_CNN_final.csv",index=False)

## 3. Split compounds into train and test set

# create cpd name - moa dictionary
df_cpds_moas = df_cellprofiler.drop_duplicates(['pert_iname','moa'])[['pert_iname','moa']]
cpds_moa = dict(zip(df_cpds_moas['pert_iname'], df_cpds_moas['moa']))
len(cpds_moa)

df_pert_cpds_moas = split_cpds_moas(cpds_moa)
df_pert_cpds_moas

len(df_pert_cpds_moas[df_pert_cpds_moas['test']]['pert_iname'].unique()) ##moas in the test data

def get_moa_count(df):
    """
    Get the number of compounds MOAs are present in, for both train and test data
    """
    df_moa_ct = df.drop(['pert_iname'], axis=1).groupby(['moa']).agg(['sum'])
    df_moa_ct.columns = df_moa_ct.columns.droplevel(1)
    df_moa_ct.reset_index(inplace=True)
    return df_moa_ct

def get_test_ratio(df):
    if df['test'] > 0:
        return df["train"] / df["test"]
    return 0

df_moa_count = get_moa_count(df_pert_cpds_moas)

df_moa_count['test_ratio'] = df_moa_count.apply(get_test_ratio, axis=1)

## All MOAs found in test should be found in train data, so this should output nothing...GOOD!
df_moa_count[(df_moa_count['train'] == 0) & (df_moa_count['test'] >= 1)]

## moas that are represented in more than one compounds (> 1), 
## present in train set but not present in test set
df_moa_count[(df_moa_count['train'] > 1) & (df_moa_count['test'] == 0)]

len(df_pert_cpds_moas[df_pert_cpds_moas['train']]['pert_iname'].unique()) ##no of compounds in train data

len(df_pert_cpds_moas[df_pert_cpds_moas['test']]['pert_iname'].unique()) ##no of compounds in test data

def save_to_csv(df, path, file_name, compress=None):
    """saves dataframes to csv"""
    
    if not os.path.exists(path):
        os.mkdir(path)
    
    df.to_csv(os.path.join(path, file_name), index=False, compression=compress)

save_to_csv(df_pert_cpds_moas, "data", 'split_moas_cpds_final.csv')
